Preserve "original order" of elements with equal weight
Previously, assigning code words to five equal-weight elements with the
alphabet "abcde" resulted in them getting the following code words,
respectively:

    "e", "d", "c", "b", "a"

This is more logical, though:

    "a", "b", "c", "d", "e"

Let's say you have a list of items in some original order. Then, that
list is sorted in ascending order, by their weights. It means that
items with lower weights go first. But what about items with _equal_
weights? If the sort is stable, they remain in their original order. It
makes sense to think that the earlier an element is in the original
order, the more important it is. This in effect means that all items in
the list are sorted by their weight in ascending order, except all items
with equal weights, which are then conceptually sorted internally in _descending_
order.

This commit fixes that problem by requiring the elements to be sorted
in _descending_ order from the beginning (instead of ascending), and
works backwards over that list. Note that this is a breaking change.